+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Add new `Rails/EnvironmentComparison` cop. ([@tdeo][])
 * Add `AllowedChars` option to `Style/AsciiComments` cop. ([@hedgesky][])
 * [#5031](https://github.com/bbatsov/rubocop/pull/5031): Add new `Style/EmptyBlockParameter` and `Style/EmptyLambdaParameter` cops. ([@pocke][])
+* [#5057](https://github.com/bbatsov/rubocop/pull/5057): Add new `Gemspec/RequiredRubyVersion` cop. ([@koic][])
 
 ### Bug fixes